# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Blog API.",
    "version": "1.0.0"
  },
  "host": "localhost:8081",
  "basePath": "/api/v1",
  "paths": {
    "/stock": {
      "get": {
        "description": "GetStocks returns all stock",
        "tags": [
          "stock"
        ],
        "operationId": "getStocks",
        "responses": {
          "200": {
            "$ref": "#/responses/Response"
          },
          "400": {
            "$ref": "#/responses/Response"
          },
          "500": {
            "$ref": "#/responses/Response"
          }
        }
      },
      "post": {
        "description": "CreateStock returns a success message",
        "tags": [
          "stock"
        ],
        "operationId": "createStock",
        "parameters": [
          {
            "x-go-name": "Body",
            "description": "Envoltura del cuerpo de la solicitud",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Stock"
            }
          }
        ],
        "responses": {
          "202": {
            "$ref": "#/responses/Response"
          },
          "400": {
            "$ref": "#/responses/Response"
          },
          "500": {
            "$ref": "#/responses/Response"
          }
        }
      }
    },
    "/stock/{id}": {
      "get": {
        "description": "GetOneStock returns one from the stock",
        "tags": [
          "stock"
        ],
        "operationId": "getStock",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "ID del stock",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Response"
          },
          "400": {
            "$ref": "#/responses/Response"
          },
          "500": {
            "$ref": "#/responses/Response"
          }
        }
      },
      "put": {
        "description": "updateStock returns a success message",
        "tags": [
          "stock"
        ],
        "operationId": "updateStock",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "ID del stock",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Body",
            "description": "Envoltura del cuerpo de la solicitud",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Stock"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Response"
          },
          "400": {
            "$ref": "#/responses/Response"
          },
          "500": {
            "$ref": "#/responses/Response"
          }
        }
      },
      "delete": {
        "description": "Delete returns message",
        "tags": [
          "stock"
        ],
        "operationId": "deleteStock",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "ID del stock",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Response"
          },
          "400": {
            "$ref": "#/responses/Response"
          },
          "500": {
            "$ref": "#/responses/Response"
          }
        }
      }
    }
  },
  "definitions": {
    "ResponseError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "Code"
        },
        "details": {
          "type": "string",
          "x-go-name": "Details"
        }
      },
      "x-go-package": "example/hello/utils"
    },
    "Stock": {
      "type": "object",
      "required": [
        "ticker",
        "target_from",
        "target_to",
        "company",
        "action",
        "brokerage",
        "rating_from",
        "rating_to"
      ],
      "properties": {
        "action": {
          "description": "Action del stock",
          "type": "string",
          "default": "upgraded by",
          "x-go-name": "Action"
        },
        "brokerage": {
          "description": "Brokerage del stock",
          "type": "string",
          "default": "The Goldman Sachs Group",
          "x-go-name": "Brokerage"
        },
        "company": {
          "description": "Company del stock",
          "type": "string",
          "default": "Banco Santander (Brasil)",
          "x-go-name": "Company"
        },
        "rating_from": {
          "description": "Rating_from del stock",
          "type": "string",
          "default": "Sell",
          "x-go-name": "Rating_from"
        },
        "rating_to": {
          "description": "Rating_to del stock",
          "type": "string",
          "default": "Neutral",
          "x-go-name": "Rating_to"
        },
        "target_from": {
          "description": "Target_from del stock",
          "type": "number",
          "format": "double",
          "default": 420,
          "minimum": 1,
          "x-go-name": "Target_from"
        },
        "target_to": {
          "description": "Target_to del stock",
          "type": "number",
          "format": "double",
          "default": 470,
          "minimum": 1,
          "x-go-name": "Target_to"
        },
        "ticker": {
          "description": "Ticker del stock",
          "type": "string",
          "default": "BSBR",
          "x-go-name": "Ticker"
        },
        "time": {
          "description": "Time del stock",
          "type": "string",
          "format": "date-time",
          "default": "2025-01-13T00:30:05.813548892Z",
          "x-go-name": "Time"
        }
      },
      "x-go-package": "example/hello/models"
    }
  },
  "responses": {
    "Response": {
      "description": "Response representa una respuesta del servidor.",
      "schema": {
        "$ref": "#/definitions/ResponseError"
      },
      "headers": {
        "data": {},
        "error": {},
        "meta": {},
        "status": {
          "type": "string"
        }
      }
    }
  }
}